+++ cli/src/index.ts
@@ -16,7 +16,7 @@ import { AstarAPI } from "./lib/api";
 import { c } from "./lib/ui";
 import { resolve } from "path";
 
-export const VERSION = "0.0.15";
+export const VERSION = "0.0.16";
 
 async function showDashboard() {
   const status = await getAuthStatus();